feat: slider functionalities implemented
made by using react-slick librarie and adjusting styles to the librarie requirements
