+++ package.json
@@ -14,6 +14,7 @@
     "next": "14.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-slick": "^0.30.2",
     "sass": "^1.77.2"
   },
   "devDependencies": {
@@ -24,6 +25,7 @@
     "@types/node": "^20.12.12",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
+    "@types/react-slick": "^0.23.13",
     "@typescript-eslint/eslint-plugin": "^7.11.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
