+++ src/app/layout.tsx
@@ -19,6 +19,19 @@ export default function RootLayout({
       className={`${fonts.lato.variable} ${fonts.absolutely_vintage.variable}`}
       lang="en"
     >
+      <head>
+        <link
+          rel="stylesheet"
+          type="text/css"
+          charSet="UTF-8"
+          href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css"
+        />
+        <link
+          rel="stylesheet"
+          type="text/css"
+          href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css"
+        />
+      </head>
       <body>
         <Header />
         {children}
