+++ src/components/FeaturedProducts/sliderFeaturedProducts.scss
@@ -1,0 +1,81 @@
+@import '../../app/globals.scss';
+
+.featuredProduct {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 10%;
+
+  width: 180px;
+  height: auto;
+  aspect-ratio: 0.77;
+  padding: 12px;
+
+  background-color: var(--c-light-violet);
+
+  @include for-tablet-portrait-up {
+    width: 250px;
+    padding: 24px;
+  }
+}
+
+.featuredProduct--image-box {
+  position: relative;
+  flex-grow: 1;
+  width: 100%;
+}
+
+.featuredProduct--image {
+  width: 100%;
+  height: auto;
+  object-fit: contain;
+}
+
+.featuredProduct--info {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  color: var(--c-ultra-violet);
+  font-size: 13px;
+  @include font-titles;
+
+  @include for-tablet-portrait-up {
+    font-size: var(--sm);
+  }
+
+  @include screen-desktop {
+    font-size: var(--md);
+  }
+}
+
+.featuredProduct--title {
+  display: inline-block;
+  max-width: 50%;
+  font-weight: 400;
+  line-height: 1.2;
+  font-size: inherit;
+}
+
+.featuredProduct--price {
+  text-align: right;
+}
+
+.slick-slider {
+  display: flex !important;
+}
+
+.slick-slide {
+  box-sizing: border-box !important;
+  transform: scale(0.9);
+  transition: all 0.4s ease-in-out;
+  height: auto;
+}
+
+.slick-slide div {
+  transition: all 0.4s ease-in-out;
+}
+
+.slick-center {
+  transform: scale(1);
+}
